+++ Task_1_1_3/src/main/java/ru/nsu/aeliseev2/task113/parsers/ExpressionParser.java
@@ -1,4 +1,4 @@
-package ru.nsu.aeliseev2.task113.parsing;
+package ru.nsu.aeliseev2.task113.parsers;
 
 import ru.nsu.aeliseev2.task113.expressions.Addition;
 import ru.nsu.aeliseev2.task113.expressions.Division;
