+++ Task_1_1_3/src/main/java/ru/nsu/aeliseev2/task113/parsers/Token.java
@@ -1,4 +1,4 @@
-package ru.nsu.aeliseev2.task113.parsing;
+package ru.nsu.aeliseev2.task113.parsers;
 
 public record Token<Data>(Data data, int position) {
 }
