+++ Task_1_1_3/src/main/java/ru/nsu/aeliseev2/task113/parsers/TokenReader.java
@@ -1,4 +1,4 @@
-package ru.nsu.aeliseev2.task113.parsing;
+package ru.nsu.aeliseev2.task113.parsers;
 
 public interface TokenReader<TokenType> extends AutoCloseable {
     Token<TokenType> peek();
