+++ Task_1_1_3/src/main/java/ru/nsu/aeliseev2/task113/parsers/UnexpectedCtxTokenException.java
@@ -1,4 +1,4 @@
-package ru.nsu.aeliseev2.task113.parsing;
+package ru.nsu.aeliseev2.task113.parsers;
 
 public class UnexpectedCtxTokenException extends ExpressionParseException {
     public final Token<EvalContextTokenData> token;
